Restructuring repository, so intersection/bounding volume traits are defined in the same place, and trait implementations are close to the type definitions
