--- src/geometry.rs
@@ -1,5 +1,0 @@
-use cgmath::BaseNum;
-
-pub trait SurfaceArea<S: BaseNum> {
-    fn surface_area(&self) -> S;
-}
