+++ src/lib.rs
@@ -48,11 +48,7 @@ pub use aabb::*;
 pub use bound::*;
 pub use cylinder::Cylinder;
 pub use frustum::*;
-pub use intersect::Continuous;
-pub use intersect::Discrete;
-pub use intersect::Contains;
-pub use geometry::SurfaceArea;
-pub use ops::Union;
+pub use traits::*;
 pub use obb::*;
 pub use sphere::Sphere;
 pub use plane::Plane;
@@ -67,11 +63,9 @@ mod aabb;
 mod bound;
 mod cylinder;
 mod frustum;
-mod intersect;
+mod traits;
 mod obb;
 mod sphere;
 mod plane;
 mod ray;
 mod line;
-mod ops;
-mod geometry;
