--- src/ops.rs
@@ -1,6 +1,0 @@
-
-pub trait Union<RHS = Self> {
-    type Output;
-
-    fn union(&self, &RHS) -> Self::Output;
-}
