+++ src/prelude.rs
@@ -1,4 +1,3 @@
-pub use intersect::{Continuous, Discrete, Contains};
-pub use ops::Union;
-pub use geometry::SurfaceArea;
+pub use traits::*;
 pub use aabb::{MinMax, Aabb};
+pub use bound::{Bound, Relation};
