+++ src/ray.rs
@@ -14,12 +14,13 @@
 // limitations under the License.
 
 use std::marker::PhantomData;
-use cgmath::{BaseFloat, BaseNum, InnerSpace};
-use cgmath::{EuclideanSpace, Point2, Point3};
-use cgmath::{VectorSpace, Vector2, Vector3};
 
-use intersect::{Continuous, Discrete};
+use cgmath::prelude::*;
+use cgmath::{BaseFloat, BaseNum};
+use cgmath::{Point2, Point3};
+use cgmath::{Vector2, Vector3};
 
+use traits::{Continuous, Discrete};
 
 /// A generic ray starting at `origin` and extending infinitely in
 /// `direction`.
